+++ CHANGELOG.md
@@ -15,12 +15,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - *NeSSI Global Optimization* references (docs/)
 - time_window function in nessi.signal windowing.py
 - space_window function in nessi.signal windowing.py
+- taper1d function in nessi.signal tapering.py
 
 ### Modified
 - nessi.swm functions are now located at nessi.modeling.swm
 - nessi.pso.Swarm class is now callable from nessi.globopt.Swarm
-- adding window in space to suwind method in nessi.io
-- suwind method now depends nessi.signal.time_window and nessi.signal.space_window
+- adding window in space to suwind method in nessi.io for SU data format
+- wind method (SU format) now depends nessi.signal.time_window and nessi.signal.space_window
+- adding taper method in nessi.io for SU data format
 
 ## [0.0.0] - 2018-05-21
 - starting point for development